MPI_Init will now detect and join a persistent universe - hooray! Fixed the session_dir cleanup process so it is kinder to the universe-setup file (i.e., leaves it alone), thus allowing persistent universes to retain their contact info on the session_dir tree. Adjusted mpirun2, ompid, and ompiconsole accordingly.
Put some error protection in ompi_rte_monitor.

This commit was SVN r2678.

+++ src/util/session_dir.c
@@ -347,7 +347,8 @@ ompi_dir_empty(char *pathname)
 		if ((0 != strcmp(ep->d_name, ".")) &&
 		    (0 != strcmp(ep->d_name, "..")) &&
 		    (DT_DIR != ep->d_type) &&
-		    (0 != strncmp(ep->d_name, "output-", strlen("output-")))) {
+		    (0 != strncmp(ep->d_name, "output-", strlen("output-"))) &&
+		    (0 != strcmp(ep->d_name, "universe-setup.txt"))) {
 		    filenm = ompi_os_path(false, pathname, ep->d_name, NULL);
 		    unlink(filenm);
 		}
